@@ -7,27 +7,34 @@
     "npm": ">=3.8"
   },
   "dependencies": {
-    "antd": "^2.13.6",
-    "axios": "^0.16.2",
+    "@antv/data-set": "^0.8.9",
+    "@nivo/bar": "^0.36.0",
+    "@nivo/line": "^0.36.0",
+    "@react-pdf/core": "^0.7.6",
+    "@react-pdf/node": "^0.7.6",
+    "ant-design-pro": "^2.0.0-beta.1",
+    "antd": "^3.5.3",
+    "axios": "^0.18.0",
     "babel-polyfill": "^6.9.1",
-    "babyparse": "^0.4.6",
+    "bizcharts": "^3.1.7",
+    "bizcharts-plugin-slider": "^2.0.3",
     "classnames": "^2.2.5",
-    "csv-parse": "^1.2.1",
+    "csv-parse": "^2.4.0",
     "fastclick": "^1.0.6",
     "fs": "0.0.1-security",
     "history": "^3.0.0",
     "jquery": "^3.2.1",
     "jquery-match-height": "^0.7.2",
+    "lodash": "^4.17.10",
+    "npm": "^5.8.0",
     "papaparse": "^4.3.5",
-    "patternfly": "^3.12.0",
-    "react": "^15.2.1",
-    "react-bootstrap-table": "^3.4.6",
-    "react-dom": "^15.2.1",
-    "react-html-parser": "^2.0.1",
+    "patternfly": "^3.44.0",
+    "prop-types": "^15.6.1",
+    "react": "^16.3.2",
+    "react-dom": "^16.3.2",
     "react-native-html-parser": "0.0.5",
-    "react-redux": "^4.4.5",
-    "redux": "^3.5.2",
-    "whatwg-fetch": "^1.0.0"
+    "react-redux": "^5.0.7",
+    "redux": "^4.0.0"
   },
   "devDependencies": {
     "assets-webpack-plugin": "^3.4.0",
@@ -35,6 +42,7 @@
     "babel-core": "^6.11.4",
     "babel-eslint": "^6.1.2",
     "babel-loader": "^6.2.4",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-runtime": "^6.9.0",
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-react": "^6.11.1",
@@ -46,14 +54,14 @@
     "connect-history-api-fallback": "^1.2.0",
     "css-loader": "^0.23.1",
     "del": "^2.2.1",
-    "ejs": "^2.5.1",
+    "ejs": "^2.5.8",
     "eslint": "^3.1.1",
     "eslint-config-airbnb": "^9.0.1",
     "eslint-plugin-import": "^1.12.0",
     "eslint-plugin-jsx-a11y": "^2.0.1",
     "eslint-plugin-react": "^5.2.2",
     "file-loader": "^0.9.0",
-    "firebase-tools": "^3.0.4",
+    "firebase-tools": "^3.18.2",
     "front-matter": "^2.1.0",
     "highlight.js": "^9.5.0",
     "json-loader": "^0.5.4",
@@ -83,7 +91,7 @@
     "url-loader": "^0.5.7",
     "webpack": "^1.13.1",
     "webpack-dev-middleware": "^1.6.1",
-    "webpack-hot-middleware": "^2.12.2"
+    "webpack-hot-middleware": "^2.22.0"
   },
   "babel": {
     "presets": [
@@ -92,7 +100,8 @@
       "stage-1"
     ],
     "plugins": [
-      "transform-runtime"
+      "transform-runtime",
+      "transform-decorators-legacy"
     ]
   },
   "eslintConfig": {
